Problem: tree min value

+++ 028-tree-min-value/breadth_first.go
@@ -1,6 +1,6 @@
 package treeminvalue
 
-func BreadthFirst(root *Node) int64 {
+func BreadthFirst(root *Node) int {
 	result := root.Val
 
 	queue := []*Node{root}
