+++ 028-tree-min-value/depth_first_recursive.go
@@ -1,6 +1,6 @@
 package treeminvalue
 
-func DepthFirstRecursive(root *Node) int64 {
+func DepthFirstRecursive(root *Node) int {
 	result := root.Val
 
 	if root.Left != nil {
